Create RWA
Relative Weights Analysis is a powerful method to gain insights into data. Very briefly, it accounts for multicollinearity between the variables and creates new independent variables which are not correlated. It then checks what percentage of the total variance (R-squared) each of these new variable accounts for.

Details about the concept and an implementation in JavaScript can be found at: https://www.listendata.com/2015/05/relative-importance-weight-analysis.html

This program assumes that the script is executed as a whole with one argument with the data as a JSON string.